+++ CHANGELOG.md
@@ -1,7 +1,23 @@
+
 x.x.x Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
 * Added templates for Expo.
+* A new mode `discardLocal` for client reset is introduce. The old behavior is supported (but deprecated) through the `manual` mode. The new mode will discard any local changes, and a fresh copy of the Realm will be downloaded. An example of the configuration:
+
+```js
+const config = {
+  // ...
+  sync: {
+    // ...
+    clientReset: {
+      mode: "discardLocal",
+      clientResyncBefore: (localRealm, remoteRealm) => { /* ... */ },
+      clientResyncAfter: (localRealm) => { /* ... */ },
+    },
+  },
+};
+```
 
 ### Fixed
 * If the option `user` in a sync configuration was not a `Realm.User` object could lead to a crash. ([#1348](https://github.com/realm/realm-js/issues/1348), since v10.0.0)